fallout v0.0.1 Fallout

Helps solve terminal hacking puzzles for Fallout 4

Functions

compare(a, b)

Find the amount of common letters at identical indexes between 2 strings

Examples

iex> Fallout.compare("WALL", "BALL")
3

count_chars(list)

Returns a map with the total character count (A-Z) of the list

Examples

iex> Fallout.count_chars(["ABC", "BCD"])
%{A: 1, B: 2, C: 2, D: 1, E: 0, F: 0, ...}

get_likeness(word)

Asks user for the likeness result and returns as an integer

hack()

Initiates a hack and asks for available words, seperated by space

hack(words)

Initiates a hack with provided words

suggest_word(word_list)

Suggests the word with most commonly used characters among the word list

vaultboy()

Prints an ASCII Vaultboy